requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==2.0.4
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.6
decorator==5.1.1
Django==5.0.6
django-allauth==65.3.0
django-appconf==1.0.6
django-bootstrap5==24.2
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-imagekit==5.0.0
django-ratelimit==4.1.0
django-sslserver==0.22
django-widget-tweaks==1.5.0
django-widgets-improved==1.5.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
ExifRead==2.3.2
Fraction==2.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
ImageHash==4.3.1
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jwt==1.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest_asyncio==1.6.0
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pilkit==3.0
pillow==10.3.0
pillow_heif==0.18.0
pip==24.3.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.7.0
pyzmq==26.0.3
ratelimit==2.2.1
requests==2.32.2
scipy==1.14.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.43.0
xlrd==2.0.1
yarl==1.13.1
zipp==3.19.2
